moved custom exceptions to a specific package
