+++ src/main/java/com/gym/gym/exceptions/customExceptions/InvalidPasswordException.java
@@ -1,4 +1,4 @@
-package com.gym.gym.exceptions;
+package com.gym.gym.exceptions.customExceptions;
 
 public class InvalidPasswordException extends RuntimeException{
     public InvalidPasswordException(String message){
