+++ src/main/java/com/gym/gym/exceptions/customExceptions/UnauthorizedAccessException.java
@@ -1,4 +1,4 @@
-package com.gym.gym.exceptions;
+package com.gym.gym.exceptions.customExceptions;
 
 public class UnauthorizedAccessException extends RuntimeException{
     public UnauthorizedAccessException(String message){
