+++ src/main/java/com/gym/gym/services/implementations/TraineeServiceImpl.java
@@ -3,7 +3,7 @@ package com.gym.gym.services.implementations;
 import com.gym.gym.entities.Trainee;
 import com.gym.gym.entities.Trainer;
 import com.gym.gym.entities.User;
-import com.gym.gym.exceptions.NotFoundException;
+import com.gym.gym.exceptions.customExceptions.NotFoundException;
 import com.gym.gym.repositories.TraineeRepository;
 import com.gym.gym.services.TraineeService;
 import com.gym.gym.services.TrainerService;
