+++ src/main/java/com/gym/gym/services/implementations/TrainerServiceImpl.java
@@ -3,7 +3,7 @@ package com.gym.gym.services.implementations;
 import com.gym.gym.entities.Trainer;
 import com.gym.gym.entities.TrainingType;
 import com.gym.gym.entities.User;
-import com.gym.gym.exceptions.NotFoundException;
+import com.gym.gym.exceptions.customExceptions.NotFoundException;
 import com.gym.gym.repositories.TrainerRepository;
 import com.gym.gym.services.TrainerService;
 import com.gym.gym.services.TrainingTypeService;
