+++ src/test/java/com/gym/gym/TraineeServiceImplTests.java
@@ -7,7 +7,7 @@ import com.gym.gym.dtos.Credentials;
 import com.gym.gym.entities.Trainee;
 import com.gym.gym.entities.Trainer;
 import com.gym.gym.entities.User;
-import com.gym.gym.exceptions.NotFoundException;
+import com.gym.gym.exceptions.customExceptions.NotFoundException;
 import com.gym.gym.repositories.TraineeRepository;
 import com.gym.gym.services.implementations.TraineeServiceImpl;
 import com.gym.gym.services.implementations.TrainerServiceImpl;
